+++ formio_data_api/models/formio_builder.py
@@ -125,31 +125,93 @@ class FormioBuilder(models.Model):
                 prop_api = comp.properties.get('server_api')
                 prop_value = comp.properties.get('server_api_value')
                 prop_value_obj = comp.properties.get('server_api_value_obj')
+                prop_value_obj_convert = comp.properties.get('server_api_value_obj_convert')
 
                 if comp_key not in data and prop_api and prop_value:
-                    component_server_api = self.component_server_api_ids.filtered(lambda x: x.active and x.name == prop_api)
-                    if not component_server_api:
+                    component_server_api_values = self.component_server_api_ids.filtered(
+                        lambda x: x.type == 'values' and x.active and x.name == prop_api
+                    )
+                    if not component_server_api_values:
                         _logger.error('NOT FOUND [formio.component.code.api] with name: %s' % prop_api)
 
-                    if component_server_api and api_values.get(component_server_api.name):
-                        value = api_values[component_server_api.name][prop_value]
+                    if component_server_api_values and api_values.get(component_server_api_values.name):
+                        value = api_values[component_server_api_values.name][prop_value]
                         if prop_value_obj:
                             # TODO-2: refactor DRY
                             value_fields = prop_value_obj.split('.')
                             value = reduce(getattr, value_fields, value)
+                            # convert value (optional feature)
+                            if prop_value_obj_convert:
+                                value = self._convert_server_api_value_obj(prop_value_obj_convert, value, comp, params)
                         data[comp_key] = value
-                    elif component_server_api:
-                        eval_context = self._get_formio_eval_context(component_server_api, self, comp, params=params)
+                    elif component_server_api_values and component_server_api_values.type == 'values':
+                        eval_context_values = self._get_formio_eval_context(component_server_api_values, self, comp, params=params)
                         # nocopy allows to return 'value'
-                        safe_eval.safe_eval(component_server_api.code, eval_context, mode="exec", nocopy=True)
-                        context_values = eval_context.get('values')
+                        safe_eval.safe_eval(component_server_api_values.code, eval_context_values, mode="exec", nocopy=True)
+                        context_values = eval_context_values.get('values')
+                        # caching
+                        if context_values:
+                            api_values[component_server_api_values.name] = context_values
+                        # value
                         value = context_values.get(prop_value)
                         if prop_value_obj:
                             # TODO-2: refactor DRY
                             value_fields = prop_value_obj.split('.')
                             value = reduce(getattr, value_fields, value)
-                        # caching
-                        if context_values:
-                            api_values[component_server_api.name] = context_values
+                        # convert value (optional feature)
+                        if prop_value_obj_convert:
+                            value = self._convert_server_api_value_obj(prop_value_obj_convert, value, comp, params)
                         data[comp_key] = value
         return data
+
+    #########
+    # convert
+    #########
+
+    def _convert_server_api_value_obj(self, component_server_api_convert_name, value, component, params={}):
+        """
+        Convert a value (in arguments) by a component API with the provided name and return result.
+
+        :param component_server_api: formio.component.server.api model record object
+        :param value: some ETL'ed (field) value, configured in the component's server_api_value_obj property
+        :param component: formiodata Component object
+        :param params: possible dict with data, eg from URL query params
+            by the /data URL endpoint
+        :returns: the component's supported type -- evaluation context given to safe_eval
+        """
+        component_server_api_convert = (
+            self.component_server_api_ids.filtered(
+                lambda x: x.type == 'convert' and x.active and x.name == component_server_api_convert_name
+            )
+        )
+        eval_context = self._get_formio_eval_context_convert_value(component_server_api_convert, value, self, component, params=params)
+        safe_eval.safe_eval(component_server_api_convert.code, eval_context, mode="exec", nocopy=True)
+        value = eval_context.get('convert')
+        return value
+
+    def _get_formio_eval_context_convert_value(self, component_server_api, value, formio_form=None, component=None, params={}):
+        """ Prepare the context used when evaluating python code
+
+        :param component_server_api: formio.component.server.api model record object
+        :param formio_form: formio.form model record object
+        :param component: formiodata Component object
+        :param params: possible dict with data, eg from URL query params
+            by the /data URL endpoint
+        :returns: the component's supported type -- evaluation context given to safe_eval
+        """
+        res = {
+            'env': self.env,
+            'builder': self,
+            'record': formio_form,
+            'form': formio_form,
+            'datetime': safe_eval.datetime,
+            'dateutil': safe_eval.dateutil,
+            'time': safe_eval.time,
+            'component': component,
+            'params': params,
+            # conversion raw input
+            'value': value,
+            # converted value
+            'convert': value
+        }
+        return res