+++ formio_data_api/models/formio_component_server_api.py
@@ -1,7 +1,8 @@
 # Copyright Nova Code (http://www.novacode.nl)
 # See LICENSE file for full licensing details.
 
-from odoo import fields, models
+from odoo import api, fields, models, _
+from odoo.exceptions import ValidationError
 
 
 class FormioComponentServerApi(models.Model):
@@ -10,6 +11,7 @@ class FormioComponentServerApi(models.Model):
 
     DEFAULT_PYTHON_CODE = """# Available variables:
 #  - values: Dictionary, read by the Forms Data API to set Components (key) value
+#  - convert: The component's supported type, read by the Forms Data API to convert and return a value
 #  - env: Odoo Environment on which the action is triggered
 #  - component: Form component (formiodata.Component) object
 #  - builder: formio.builder object (record) on which the action is triggered
@@ -24,7 +26,7 @@ class FormioComponentServerApi(models.Model):
     name = fields.Char(string='API Name', required=True)
     type = fields.Selection(
         string='Type',
-        selection=[('values', 'values')],
+        selection=[('values', 'values'), ('convert', 'convert')],
         default='values',
         ondelete='set null',
     )
@@ -36,3 +38,10 @@ class FormioComponentServerApi(models.Model):
         help="Write Python code that the Component Value will get set. Some variables are "
         "available for use; help about python expression is given in the help tab.")
     active = fields.Boolean(string='Active', default=True)
+
+    @api.constrains('name', 'type')
+    def _constraint_unique(self):
+        domain = [('name', '=', self.name), ('type', '=', self.type)]
+        if self.search(domain):
+            msg = _('Components API Name and Type should be unique.\nOnly 1 Component API with name "%s" and type "%s" is allowed.') % (self.name, self.type)
+            raise ValidationError(msg)